+++ src/functions/httpGetScrapedData.ts
@@ -1,1 +1,1 @@
-import { app, HttpRequest, HttpResponseInit, InvocationContext } from '@azure/functions';import { DataService } from '../services/dataService';export async function httpGetScrapedData (request: HttpRequest, context: InvocationContext): Promise<HttpResponseInit> {  const data = await new DataService().getData();  return {    status: 200,    headers: {      'Content-Type': 'application/json',    },    body: JSON.stringify(data),  };}app.http('httpGetScrapedData', {  methods: ['GET'],  authLevel: 'anonymous',  handler: httpGetScrapedData,});
+import { app, HttpRequest, HttpResponseInit, InvocationContext } from '@azure/functions';import { getData } from '../data/dataRepository';export async function httpGetScrapedData (request: HttpRequest, context: InvocationContext): Promise<HttpResponseInit> {  const data = await getData();  return {    status: 200,    headers: {      'Content-Type': 'application/json',    },    body: JSON.stringify(data),  };}app.http('httpGetScrapedData', {  methods: ['GET'],  authLevel: 'anonymous',  handler: httpGetScrapedData,});
