--- src/types.ts
@@ -1,18 +1,0 @@
-interface Data {
-    marketplaces: string[],
-    brands: Brand[];
-}
-
-interface Brand {
-    name: string,
-    products: Product[];
-}
-
-interface Product {
-    marketplace: string,
-    title: string,
-    currentPrice: string,
-    oldPrice: string | null,
-    imgSrc: string | null,
-    volume: string | null;
-}
